+++ collector/src/main/java/com/bitmovin/analytics/stateMachines/PlayerStateMachine.kt
@@ -38,6 +38,7 @@ class PlayerStateMachine(
     private var elapsedTimeOnEnter: Long = 0
 
     var isStartupFinished = false
+    private var isProgramChange = false
 
     var videoTimeStart: Long = 0
         private set
@@ -168,8 +169,10 @@ class PlayerStateMachine(
         videoTimeStart = videoTimeEnd
     }
 
-    private fun resetSourceRelatedState() {
-        disableHeartbeat()
+    private fun resetSourceRelatedState(keepHeartbeatRunning: Boolean = false) {
+        if (!keepHeartbeatRunning) {
+            disableHeartbeat()
+        }
         disableRebufferHeartbeat()
         videoStartTimeoutTimer.cancel()
         bufferingTimeoutTimer.cancel()
@@ -278,6 +281,33 @@ class PlayerStateMachine(
         if (shouldStartup) {
             transitionState(PlayerStates.STARTUP, newVideoTime, null)
         }
+    }
+
+    fun programChange(
+        videoTime: Long,
+        onAfterSessionReset: () -> Unit,
+    ) {
+        val lastStateBeforeProgramChange = currentState
+        triggerLastSampleOfSession()
+        resetSourceRelatedState(keepHeartbeatRunning = true)
+        onAfterSessionReset()
+
+        isProgramChange = true
+        isStartupFinished = true
+        currentState = PlayerStates.STARTUP
+
+        // If player is currently playing, transition back to playing
+        // We don't transition into Playing if the player is seeking
+        // This check is needed because some players (like THEO) return isPlaying=true while seeking
+        if (playerContext.isPlaying() && lastStateBeforeProgramChange != PlayerStates.SEEKING) {
+            transitionState(PlayerStates.PLAYING, videoTime, null)
+        }
+    }
+
+    fun getAndResetIsProgramChange(): Boolean {
+        val value = isProgramChange
+        isProgramChange = false
+        return value
     }
 
     fun pause(position: Long) {